refact: split search lecture list, make search lecture cell stateless
